requirements.txt
langchain
faiss-cpu
sqlite-utils
python-dotenv
groq